Refactor logging setup to use consistent module paths and names.
